Move MapboxVectorLayer over from ol
